evdev: fix aspect ratio when using fit/fill
The code before this commit called remap by translating output width
into wacom coordinates, so that remap returns wacom coordinates.
But this does not work correctly with the aspect ratio calculation that
remap is doing: in wacom coordinates, 1 unit in x direction may not be
the same length as 1 unit in y direction (since wacom coordinates get
stretched to the full screen by the input driver).

This commit fixes the issue by calling remap with proper output widths,
and then translating the result back into wacom coordinates for
rewriting the event.